requirements.txt
django>=4.2,<5.0
djangorestframework
django-cors-headers
flake8==5.0.4
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
pytest==7.1.3
pytest-django==4.5.2
Pillow
python-dotenv
djangorestframework-simplejwt
drf-spectacular